Fix SSR warning emitted by gatsby develop (#1)
* Fix SSR warning emitted by gatsby develop

This error message appears in the `gatsby develop` console wherever `useIsJsEnabled` is used.
```
Warning: useLayoutEffect does nothing on the server, because its effect cannot be encoded into the server renderer's output format. This will lead to a mismatch between the initial, non-hydrated UI and the intended UI. To avoid this, useLayoutEffect should only be used in components that render exclusively on the client. See https://reactjs.org/link/uselayouteffect-ssr for common fixes.
```

* Change to useEffect() only